Add support for `AdvanceErrorHandling` when sending a message
In addition to the support of the new field, this commit improves the
testing of the SendMailV31 method. The tests now validate that the
request payload matches the passed input and that encoding and
authorization information are present. The APIErrorDetailsV31 struct now
also supports the ErrorCode and ErrorIdentifier fields. The ErrorClass
field of that struct is now deprecated.

Signed-off-by: Marius Kießling <marius.kiessling@metro.digital>